target/vgmini/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stock-trading-ml-framework"
version = "1.0.0"
description = "ML framework for stock trading signal generation and backtesting"
authors = [
    {name = "Trading Framework", email = "framework@example.com"},
]
dependencies = [
    "pyyaml>=6.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "psycopg2-binary>=2.9.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "shap>=0.46.0",
    "ta>=0.10.0",
    "plotly>=5.15.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "flask>=3.0.0",
    "numba>=0.58.0",
    "pyarrow>=14.0.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
trading-framework = "main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.hatch.build.targets.wheel]
packages = ["src"]
